+++ ocrd_models/ocrd_models/ocrd_exif.py
@@ -3,6 +3,10 @@ Technical image metadata
 """
 
 from math import sqrt
+from io import BytesIO
+from subprocess import run, PIPE
+from distutils.spawn import find_executable as which
+from ocrd_utils import getLogger
 
 class OcrdExif():
     """Represents technical image metadata.
@@ -34,33 +38,31 @@ class OcrdExif():
         self.height = img.height
         self.photometricInterpretation = img.mode
         self.n_frames = img.n_frames if 'n_frames' in img.__dict__ else 1
-        #  if img.format == 'PNG':
-        #      print(img.info)
         for prop in ['compression', 'photometric_interpretation']:
             setattr(self, prop, img.info[prop] if prop in img.info else None)
-        if img.format in ('TIFF', 'PNG') and 'dpi' in img.info:
-            self.xResolution = int(img.info['dpi'][0])
-            self.yResolution = int(img.info['dpi'][1])
-            if img.format == 'TIFF':
-                self.resolutionUnit = 'cm' if img.tag.get(296) == 3 else 'inches'
+        if img.filename:
+            ret = run(['identify', '-format', r'%[resolution.x] %[resolution.y] %U', img.filename], check=False, stderr=PIPE, stdout=PIPE)
+        else:
+            with BytesIO() as bio:
+                img.save(bio, format=img.format)
+                ret = run(['identify', '-format', r'%[resolution.x] %[resolution.y] %U', '/dev/stdin'], check=False, stderr=PIPE, stdout=PIPE, input=bio.getvalue())
+        if ret.returncode:
+            stderr = ret.stderr.decode('utf-8')
+            if not which('identify'):
+                raise Exception("The 'identify' command is not available. Install with 'sudo apt install imagemagick'")
+            if 'no decode delegate for this image format' in stderr:
+                getLogger('ocrd_exif').warning("ImageMagick does not support the '%s' image format. ", img.format)
             else:
-                self.resolutionUnit = 'inches'
-        elif img.format == 'JPEG' and 'jfif_density' in img.info:
-            self.xResolution = img.info['jfif_density'][0]
-            self.yResolution = img.info['jfif_density'][1]
-            self.resolutionUnit = 'cm' if img.info['jfif_unit'] == 2 else 'inches'
-        elif img.format == 'PNG' and 'aspect' in img.info:
-            self.xResolution = img.info['aspect'][0]
-            self.yResolution = img.info['aspect'][1]
+                getLogger('ocrd_exif').error("identify exited with non-zero %s: %s", ret.returncode, stderr)
+            self.xResolution = self.yResolution = 1
             self.resolutionUnit = 'inches'
         else:
-            #  if img.format == 'JPEG2000':
-            #      import sys
-            #      print('JPEG 2000 not supported yet :(', file=sys.stderr)
-            self.xResolution = 1
-            self.yResolution = 1
-            self.resolutionUnit = 'inches'
-        #  print('format=%s type=%s' % (img.format, type(self.xResolution))
+            tokens = ret.stdout.decode('utf-8').split(' ', 3)
+            self.xResolution = max(int(float(tokens[0])), 1)
+            self.yResolution = max(int(float(tokens[1])), 1)
+            self.resolutionUnit = 'inches' if tokens[2] == 'undefined' else \
+                                  'cm' if tokens[2] == 'PixelsPerCentimeter' else \
+                                  'inches'
         self.resolution = round(sqrt(self.xResolution * self.yResolution))
 
     def to_xml(self):