+++ tests/test_resolver.py
@@ -9,16 +9,22 @@ from pathlib import (
 from unittest import (
     mock
 )
+from PIL import (
+    Image
+)
+
+from ocrd_models.ocrd_page import OcrdPage
 
 import pytest
 
 from tests.base import (
-    assets, 
+    assets,
     main
 )
 
 from ocrd.resolver import Resolver
 from ocrd_utils import pushd_popd
+
 
 # set pylint once on module level
 # pylint: disable=protected-access
@@ -73,8 +79,7 @@ def test_workspace_from_url_kant(mock_request, tmp_path):
 
     # act
     resolver = Resolver()
-    resolver.workspace_from_url(
-        url_src, mets_basename='foo.xml', dst_dir=dst_dir)
+    resolver.workspace_from_url(url_src, mets_basename='foo.xml', dst_dir=dst_dir)
 
     # assert
     local_path = dst_dir / 'foo.xml'
@@ -94,11 +99,7 @@ def test_workspace_from_url_kant_with_resources(mock_request, tmp_path):
 
     # act
     resolver = Resolver()
-    resolver.workspace_from_url(
-        url_src,
-        mets_basename='kant_aufklaerung_1784.xml',
-        dst_dir=dst_dir,
-        download=True)
+    resolver.workspace_from_url(url_src, mets_basename='kant_aufklaerung_1784.xml', dst_dir=dst_dir, download=True)
 
     # assert files present under local tmp_path
     local_path_mets = dst_dir / 'kant_aufklaerung_1784.xml'
@@ -120,15 +121,12 @@ def test_workspace_from_url_kant_with_resources_existing_local(mock_request, tmp
     mock_request.side_effect = request_behavior
     dst_dir = tmp_path / 'workspace_kant'
     dst_dir.mkdir()
-    src_mets = Path(assets.path_to(
-        'kant_aufklaerung_1784-binarized/data/mets.xml'))
+    src_mets = Path(assets.path_to('kant_aufklaerung_1784-binarized/data/mets.xml'))
     dst_mets = Path(dst_dir, 'mets.xml')
     shutil.copyfile(src_mets, dst_mets)
 
     # act
-    Resolver().workspace_from_url(url_src,
-                                  clobber_mets=False,
-                                  dst_dir=dst_dir)
+    Resolver().workspace_from_url(url_src, clobber_mets=False, dst_dir=dst_dir)
 
     # assert
     # no real request was made, since mets already present
@@ -153,8 +151,7 @@ def test_workspace_from_url_404(mock_request):
 
 
 def test_workspace_from_url_with_rel_dir(tmp_path):
-    bogus_dst_dir = '../../../../../../../../../../../../../../../../%s' % str(tmp_path)[
-        1:]
+    bogus_dst_dir = '../../../../../../../../../../../../../../../../%s' % str(tmp_path)[1:]
 
     # act
     with pushd_popd(FOLDER_KANT):
@@ -189,19 +186,26 @@ def test_resolve_image0():
 
 
 @pytest.mark.parametrize(
-    "image_url,size1,size2",
-    [('OCR-D-IMG-NRM/OCR-D-IMG-NRM_0017.png', (1457, 2083), (1, 1)),
-     ('OCR-D-IMG-1BIT/OCR-D-IMG-1BIT_0017.png', (1457, 2083), (1, 1)),
+    "image_url,size_pil",
+    [('OCR-D-IMG-NRM/OCR-D-IMG-NRM_0017.png', (1, 1)),
+     ('OCR-D-IMG-1BIT/OCR-D-IMG-1BIT_0017.png', (1, 1)),
      ])
-def test_resolve_image_grayscale(image_url, size1, size2):
-    url_path = os.path.join(assets.url_of(
-        'kant_aufklaerung_1784-binarized'), 'data/mets.xml')
+def test_resolve_image_as_pil(image_url, size_pil):
+    url_path = assets.url_of('kant_aufklaerung_1784-binarized/data/mets.xml')
     workspace = Resolver().workspace_from_url(url_path)
-    img_url = image_url
-    img_pil1 = workspace.resolve_image_as_pil(img_url)
-    assert img_pil1.size == size1
-    img_pil2 = workspace._resolve_image_as_pil(img_url, [[0, 0], [1, 1]])
-    assert img_pil2.size == size2
+    img_pil = workspace._resolve_image_as_pil(image_url, [[0, 0], [1, 1]])
+    assert img_pil.size == size_pil
+
+
+def test_resolve_image_as_pil_deprecated():
+    url_path = os.path.join(assets.url_of('kant_aufklaerung_1784-binarized'), 'data/mets.xml')
+    workspace = Resolver().workspace_from_url(url_path)
+    with pytest.warns(DeprecationWarning) as record:
+        workspace.resolve_image_as_pil('OCR-D-IMG-NRM/OCR-D-IMG-NRM_0017.png')
+
+    # assert
+    assert len(record) == 1
+    assert 'Call to deprecated method resolve_image_as_pil.' in str(record[0].message)
 
 
 def test_workspace_from_nothing():
@@ -230,11 +234,10 @@ def test_workspace_from_nothing_noclobber(tmp_path):
     assert the_msg in str(exc)
 
 
-@pytest.mark.parametrize(
-    "url,basename,exc_msg",
-    [(None, None, "'url' must be a string"),
-     (None, 'foo', "'directory' must be a string")]
-)
+@pytest.mark.parametrize("url,basename,exc_msg",
+                         [(None, None, "'url' must be a string"),
+                          (None, 'foo', "'directory' must be a string")]
+                         )
 def test_download_to_directory_with_badargs(url, basename, exc_msg):
 
     with pytest.raises(Exception) as exc:
@@ -262,8 +265,7 @@ def test_download_to_directory_default(fixture_copy_kant):
 def test_download_to_directory_basename(fixture_copy_kant):
     tmp_root = fixture_copy_kant.parent
     phil_data = fixture_copy_kant / 'data' / 'mets.xml'
-    fn = Resolver().download_to_directory(
-        str(tmp_root), str(phil_data), basename='foo')
+    fn = Resolver().download_to_directory(str(tmp_root), str(phil_data), basename='foo')
     assert Path(tmp_root, fn).exists()
     assert fn == 'foo'
 
@@ -271,8 +273,7 @@ def test_download_to_directory_basename(fixture_copy_kant):
 def test_download_to_directory_subdir(fixture_copy_kant):
     tmp_root = fixture_copy_kant.parent
     phil_data = fixture_copy_kant / 'data' / 'mets.xml'
-    fn = Resolver().download_to_directory(
-        str(tmp_root), str(phil_data), subdir='baz')
+    fn = Resolver().download_to_directory(str(tmp_root), str(phil_data), subdir='baz')
     assert Path(tmp_root, fn).exists()
     assert fn == 'baz/mets.xml'